+++ phpstan-bootstrap.php
@@ -9,5 +9,9 @@
 define( 'SWH_PLUGIN_URL', '' );
 define( 'SWH_PLUGIN_DIR', '' );
 define( 'SWH_PLUGIN_FILE', '' );
-define( 'SWH_VERSION', '2.1.0' );
+define( 'SWH_VERSION', '2.2.0' );
+define( 'SWH_CDN_BASE', 'https://media.seanmousseau.com/file/seanmousseau/assets/logos/swh' );
+define( 'SWH_ICON_1X', SWH_CDN_BASE . '/icon-128x128.png' );
+define( 'SWH_ICON_2X', SWH_CDN_BASE . '/icon-256x256.png' );
+define( 'SWH_MENU_ICON', SWH_ICON_1X );